refactor: add SetLoaderTest trait for RoutesLoader and ObjectivesLoader, and rename package 'utils' to 'config'
